refactor: container 추가, container 추가에 따른 변경사항 수정 및 type, utils 분리
