+++ durihana/app/globals.css
@@ -224,8 +224,7 @@
     min-height: 100dvh;
     min-height: -webkit-fill-available;
     min-height: 100vh;
-    overflow-x: hidden;
-    overscroll-behavior: none;
+    overflow: hidden;
     font-family: var(--font-hana);
     /* 스크롤바 제거 */
     -ms-overflow-style: none; /* IE & Edge */
