+++ durihana/constants/store.ts
@@ -1,0 +1,17 @@
+export const categories = [
+  { id: 1, name: '예식장', icon: '/asset/icons/flower.svg' },
+  { id: 2, name: '스드메', icon: '/asset/icons/dress.svg' },
+  { id: 3, name: '신혼여행', icon: '/asset/icons/flight.svg' },
+  { id: 4, name: '가전·가구', icon: '/asset/icons/tv.svg' },
+  { id: 5, name: '예물·예단', icon: '/asset/icons/ring.svg' },
+];
+
+export const partners = [
+  { label: '예식장', img: '/asset/images/weddinghall-popular.png' },
+  { label: '신혼여행', img: '/asset/images/tour-popular.png' },
+  { label: '스튜디오', img: '/asset/images/studio-popular.png' },
+  { label: '드레스', img: '/asset/images/dress-popular.png' },
+  { label: '가전', img: '/asset/images/electronic-popular.png' },
+  { label: '가구', img: '/asset/images/furniture-popular.png' },
+  { label: '예물', img: '/asset/images/ring-popular.png' },
+];
