+++ durihana/lib/actions/DashboardActions.ts
@@ -3,6 +3,7 @@
 import prisma from '../db';
 
 export const getMarriageDate = async (userId: number) => {
+  if (!userId) return '';
   const user = await prisma.user.findUnique({
     where: { id: userId },
   });
@@ -11,10 +12,7 @@ export const getMarriageDate = async (userId: number) => {
 };
 
 export const getCategoriesByUserId = async (userId?: number) => {
-  if (!userId) {
-    return [];
-  }
-
+  if (!userId) return [];
   const calendars = await prisma.partnerCalendar.findMany({
     where: { user_id: userId },
     select: {
